Refactor project structure and migrate to `src` directory
Moved all modules and files to a unified `src` directory for better organization and maintainability. Updated import paths across the project to reflect these changes. Removed unused `database_layer.facade` module as it was redundant.
