+++ .gitignore
@@ -1,6 +1,6 @@
 /.idea/csv-editor.xml
-/Web_Layer/templates/mark_points.html
-/Web_Layer/templates/mark_points_layer.html
+/src/Web_Layer/templates/mark_points.html
+/src/Web_Layer/templates/mark_points_layer.html
 /.idea/misc.xml
 /uv.lock
 /*.iml
