+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "requests>=2.32.3",
     "shiny[all]>=1.4.0",
     "shinywidgets>=0.5.2",
+    "sphinx>=8.2.3",
     "sqlalchemy>=2.0.40",
     "uv>=0.6.17",
     "uvicorn>=0.34.2",
