+++ src/Web_Layer/templates/mark_points_from_city.html
