+++ src/database_layer/csv_to_address.py
@@ -1,8 +1,8 @@
 import asyncio
 import csv
 
-from database_layer.db_service import DBService
-from domain.DatabaseModelClasses import Address
+from src.database_layer.db_service import DBService
+from src.domain import Address
 
 
 class DataReader:
